// frontend/src/features/chat/ui/chat-entity-details/__generated__/channelContentUpdateMetadataMutation.graphql.ts
export type channelContentUpdateMetadataMutation$variables = {
  id: string;
  slug?: string | null | undefined;
  title?: string | null | undefined;
};
export type channelContentUpdateMetadataMutation$data = {
  readonly updateChatMetadata: {
    readonly __typename: "Chat";
    readonly id: string;
    readonly slug: string | null | undefined;
    readonly title: string;
  } | {
    // This will never be '%other', but we need some
    // value in case none of the concrete values match.
    readonly __typename: "%other";
  };
};
export type channelContentUpdateMetadataMutation = {
  response: channelContentUpdateMetadataMutation$data;
  variables: channelContentUpdateMetadataMutation$variables;
};